fix soft-lock, add soft/hard reset
Previous code set a flag when air in line detected, this would prevent the code from calling .readSample(), in the main loop. as readSample is the function that updates the I2C buffer, the code would have soft-locked when air was detected in-line, this is because when the airInLineDetected() function is called, the function only accesses the most recently read sample from the I2C buffer, which would never update. 

Added the ability for the code to soft-reset the I2C bus if communication fails, and hard reset (restart MCU) if soft reseting does not reestablish connection within 0.5 s.